refactor(msg-gen): move code into generator helper
