+++ packages/engine_umbrella/apps/message_generator/lib/message_generator.ex
@@ -1,13 +1,9 @@
 defmodule MessageGenerator do
-  alias MessageGenerator.Address
   alias MessageGenerator.Adapters.RMQ
+  alias MessageGenerator.Generator
 
   @transports_exchange "incoming_vehicle_updates"
   @bookings_exchange "incoming_booking_updates"
-
-  @stockholm %{lat: 59.3414072, lon: 18.0470482}
-  @gothenburg %{lat: 57.7009147, lon: 11.7537571}
-  @ljusdal %{lat: 61.829182, lon: 16.0896213}
 
   def add_random_transport(properties \\ %{})
 
@@ -18,30 +14,20 @@ defmodule MessageGenerator do
 
   def add_random_transport(city) when city in [:stockholm, :gothenburg] do
     %{}
-    |> add_transport_addresses(city)
+    |> Generator.put_new_transport_addresses_from_city(city)
     |> generate_transport()
     |> RMQ.publish(@transports_exchange, "registered")
   end
 
   def generate_transport(properties \\ %{}) do
     properties
-    |> Map.put_new(:start_address, Address.random(@ljusdal))
-  end
-
-  def add_transport_addresses(map), do: do_add_transport_addresses(map, @ljusdal)
-  def add_transport_addresses(map, :stockholm), do: do_add_transport_addresses(map, @stockholm)
-  def add_transport_addresses(map, :gothenburg), do: do_add_transport_addresses(map, @gothenburg)
-
-  defp do_add_transport_addresses(map, location) do
-    map
-    |> Map.put(:start_address, Address.random(location))
-    |> Map.put(:end_address, Address.random(location))
+    |> Generator.put_new_transport_addresses_from_city(:ljusdal)
   end
 
   def generate_booking(properties \\ %{}) do
     properties
     |> Map.put_new(:externalId, Enum.random(0..100_000))
-    |> add_booking_addresses()
+    |> Generator.add_booking_addresses()
     |> Map.put_new(:size, %{measurements: [105, 55, 26], weight: Enum.random(1..200)})
     |> Map.put_new(:metadata, %{
       sender: %{contact: "0701234567"},
@@ -58,18 +44,8 @@ defmodule MessageGenerator do
 
   def add_random_booking(city) when city in [:stockholm, :gothenburg] do
     %{}
-    |> add_booking_addresses(city)
+    |> Generator.add_booking_addresses(city)
     |> generate_booking()
     |> RMQ.publish(@bookings_exchange, "registered")
-  end
-
-  def add_booking_addresses(map), do: do_add_booking_addresses(map, @ljusdal)
-  def add_booking_addresses(map, :stockholm), do: do_add_booking_addresses(map, @stockholm)
-  def add_booking_addresses(map, :gothenburg), do: do_add_booking_addresses(map, @gothenburg)
-
-  defp do_add_booking_addresses(map, location) do
-    map
-    |> Map.put(:pickup, Address.random(location))
-    |> Map.put(:delivery, Address.random(location))
   end
 end
